Extract precaution_relevance.rs from turn.rs (parent #680)
Create new sibling module `src/agent/loop_run/precaution_relevance.rs`
to host the Issue #453 per-prompt precaution selection pipeline (sort
by severity + relevance score + budget caps). Small follow-on
extraction in the meta-issue #680 reduction effort.

Migrated to `precaution_relevance.rs`:

* `pub fn select_precautions_for_prompt(active_precautions, mode,
    touched_files, suspected_files) -> Vec<Precaution>` — entry point,
    re-exported via `loop_run.rs::pub use precaution_relevance::...`
    so `commands.rs` + `tests/precaution_prompt_injection.rs` +
    `tests/precautions_repl.rs` keep their existing import paths.
* `pub(super) fn normalize_relevance_key(s) -> String`
* `pub(super) fn relevance_keyset_from_touched(items) -> HashSet<String>`
* `pub(super) fn relevance_keyset_from_suspected(paths) -> HashSet<String>`
* `pub(super) fn relevance_score(&Precaution, &HashSet, &HashSet) -> u8`
    (Codex CB-001 fix order: suspected > touched > global > unrelated)
* `pub(super) fn sort_precautions_for_prompt(active, touched, suspected)
    -> Vec<&Precaution>` (stable sort by severity then relevance)
* `pub(super) fn apply_budget_caps(sorted) -> Vec<Precaution>` (hard
    cap on count + soft cap on cumulative `chars()`, bypassed once
    for the first oversized item per DR1-002)

`loop_run.rs` adjustments:

* Registered `mod precaution_relevance;` with a parent-#680 doc
  comment.
* Updated the existing `pub use turn::select_precautions_for_prompt`
  to `pub use precaution_relevance::select_precautions_for_prompt`
  so external import paths (`anvil::agent::loop_run::*`) are
  unchanged.

`turn.rs` adjustments:

* Removed the 7 fn definitions (originally at lines 794-954).
* Added `use super::precaution_relevance::select_precautions_for_prompt`
  for the production caller.
* Test mod imports rewrite: `use super::{apply_budget_caps, ...}` →
  `use super::super::precaution_relevance::{...}`.
* Pruned now-unused file-level imports: `std::collections::HashSet`,
  `WorkingMemory` (from `crate::session::store::{...}`),
  `Precaution` / `severity_order` (from
  `crate::session::precaution::{...}`).

LOC delta:

* `turn.rs`: 28,395 → 28,234 (-161 LOC)
* `precaution_relevance.rs`: 0 → 184 (new)
* Cumulative `turn.rs`: 39,489 → 28,234 (-11,255 LOC, -28.5%)

No facade re-export beyond the pre-existing `select_precautions_for_prompt`
public surface (DR3-001).

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
